+++ service/src/main/java/com/dashnet/dashNet/User/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 	List<User> findAll();
-	User findUserById(Integer id);
+	User findUserById(Long id);
 
 	User findUserByEmail(String email);
 }
